+++ R/crs.R
@@ -251,7 +251,7 @@ fm_crs_set_ellipsoid_radius <- function(crs, radius) {
 #' `fm_wkt_unit_params()`
 #' @returns For `fm_wkt_unit_params`, a
 #' list of named unit definitions
-#' @author Finn Lindgren \email{finn.lindgren@@gmail.com}
+#' @author Finn Lindgren <Finn.Lindgren@@gmail.com>
 #' @examples
 #' \donttest{
 #' c1 <- fm_crs("globe")
@@ -550,7 +550,7 @@ fm_length_unit.character <- function(x) {
 #'
 #' @returns A `crs` object ([sf::st_crs()]) or a `fm_crs` object.
 #' An S3 `fm_crs` object is a list with elements `crs` and `oblique`.
-#' @author Finn Lindgren \email{finn.lindgren@@gmail.com}
+#' @author Finn Lindgren <Finn.Lindgren@@gmail.com>
 #' @seealso [sf::st_crs()], [`fm_crs_wkt`]
 #' @examples
 #' crs1 <- fm_crs("longlat_globe")
@@ -585,7 +585,7 @@ fm_crs <- function(x, oblique = NULL, ..., crsonly = deprecated()) {
       )
     }
 
-    lifecycle::deprecate_warn(
+    lifecycle::deprecate_stop(
       "0.0.1",
       "fm_crs(crsonly=' should no longer be used')",
       "fm_crs(oblique)",
@@ -1127,7 +1127,7 @@ fm_crs.matrix <- function(x, oblique = NULL, ...) {
 #' An S3 `inla.CRS` object is a list, usually (but not necessarily)
 #' containing at least one element: \item{crs }{The basic `sp::CRS`
 #' object}
-#' @author Finn Lindgren \email{finn.lindgren@@gmail.com}
+#' @author Finn Lindgren <Finn.Lindgren@@gmail.com>
 #' @seealso [fm_crs()], [sp::CRS()], [`fm_crs_wkt`],
 #' [fm_sp_get_crs()], [fm_crs_is_identical()]
 #' @examples
@@ -1713,7 +1713,7 @@ fm_list_as_CRS <- function(x, ...) {
 #' name/value pairs.
 #'
 #' For `fm_list_as_CRS`, a `CRS` or `inla.CRS` object.
-#' @author Finn Lindgren <finn.lindgren@@gmail.com>
+#' @author Finn Lindgren <Finn.Lindgren@@gmail.com>
 #' @seealso [fm_CRS()]
 #' @export
 #' @keywords internal
@@ -2137,7 +2137,7 @@ fm_crs_is_identical <- function(crs0, crs1, crsonly = FALSE) {
 #' by `fm_crs_is_identical()`.
 #' @export
 fm_identical_CRS <- function(crs0, crs1, crsonly = FALSE) {
-  lifecycle::deprecate_warn(
+  lifecycle::deprecate_stop(
     "0.1.0",
     "fm_identical_CRS()",
     "fm_crs_is_identical()"
@@ -2767,7 +2767,7 @@ fm_has_PROJ6 <- function() {
 #' `sp::Spatial` and `sp::CRS` objects.
 #' @export
 fm_as_sp_crs <- function(x, ...) {
-  lifecycle::deprecate_warn(
+  lifecycle::deprecate_stop(
     "0.0.1",
     "fm_as_sp_crs()",
     "fm_CRS()"
@@ -2782,7 +2782,7 @@ fm_as_sp_crs <- function(x, ...) {
 #'   `CRS(wkt)` for `sp::Spatial` objects.
 #' @param x A `sp::Spatial` object
 #' @returns A `CRS` object, or NULL if no valid CRS identified
-#' @author Finn Lindgren \email{finn.lindgren@@gmail.com}
+#' @author Finn Lindgren <Finn.Lindgren@@gmail.com>
 #' @details This function is a convenience method to workaround PROJ4/PROJ6
 #'   differences, and the lack of a crs extraction method for Spatial objects.
 #'   For newer code, use [fm_crs()] instead, that returns `crs` objects, and use